miquelangel to main

+++ public/PCEconomic/src/main/java/me/pceconomic/shop/domain/entities/article/Valoracions.java
@@ -16,7 +16,7 @@ public @Data class Valoracions {
     private int id;
 
     @Column(name = "valoracio")
-    private int valoracio;
+    private double valoracio;
 
     @Column(name = "comentari")
     private String comentari;
